feat: support fractional percentages for disk.min_free_percent
- Change type from u8 to f64
- Add validation: 0.0 <= min_free_percent <= 100.0
- Update disk_monitor to use f64 directly
- Fix test literals to use float (99.0, 1.0)
- Add unit tests for validation and TOML parsing (0.1)
- Update config.example.toml with comment about fractional support

+++ src/agent/disk_monitor.rs
@@ -217,7 +217,7 @@ impl DiskMonitor {
             Ok(available_bytes) => match fs2::total_space(&path) {
                 Ok(total_bytes) => {
                     let available_percent = (available_bytes as f64 / total_bytes as f64) * 100.0;
-                    let threshold = self.config.min_free_percent as f64;
+                    let threshold = self.config.min_free_percent;
 
                     let currently_degraded = self.is_degraded();
